bookeater/backend/requirements.txt
dj-rest-auth==2.1.3
Django==3.2
django-allauth==0.47.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
oauthlib==3.1.1
requests-oauthlib==1.3.0
Pillow==9.2.0
six==1.16.0
gunicorn
psycopg2-binary==2.9.3
